(Admin) Controller Mountain, HikingRoute OK minus pencarian

+++ app/Http/Controllers/admin/AdminHikingRouteController.php
@@ -36,14 +36,15 @@ class AdminHikingRouteController extends Controller
                 'location' => 'required|string',
                 'distance' => 'required|integer',
                 'duration' => 'required|integer',
-                'elevation_gain' => 'required|integer',
                 'operating_hours' => 'nullable|string',
                 'numb_of_posts' => 'required|integer',
                 'contact' => 'nullable|string',
                 'fee' => 'nullable|integer',
                 'img' => 'nullable',
-                'link' => 'nullable',
                 'rules' => 'nullable',
+                'lat' => 'required',
+                'lng' => 'required',
+                'file' => 'nullable',
             ]
         );
 
@@ -51,11 +52,7 @@ class AdminHikingRouteController extends Controller
             return redirect()->back()->withErrors($validator)->withInput();
         }
 
-        // Menangani gambar jika ada
-        $img = $request->file('img');
-        $img->storeAs('public/hikingroutes', $img->hashName());
-
-        HikingRoute::create([
+        $data = [
             'mountain_id' => $request->mountain_id,
             'name' => $request->name,
             'status' => $request->status,
@@ -63,15 +60,28 @@ class AdminHikingRouteController extends Controller
             'location' => $request->location,
             'distance' => $request->distance,
             'duration' => $request->duration,
-            'elevation_gain' => $request->elevation_gain,
             'operating_hours' => $request->operating_hours,
             'numb_of_posts' => $request->numb_of_posts,
             'contact' => $request->contact,
             'fee' => $request->fee,
-            'img' => $img->hashName(),
-            'link' => $request->link,
             'rules' => $request->rules,
-        ]);
+            'lat' => $request->lat,
+            'lng' => $request->lng,
+        ];
+
+        if ($request->hasFile('img')) {
+            $img = $request->file('img');
+            $img->storeAs('public/hikingroutes', $img->hashName());
+            $data['img'] = $img->hashName();
+        }
+
+        if ($request->hasFile('file')) {
+            $file = $request->file('file');
+            $file->storeAs('public/gpx', $file->hashName());
+            $data['file'] = $file->hashName();
+        }
+
+        HikingRoute::create($data);
 
         return redirect()->route('admin.hikingroutes.index')->with('success', 'Data berhasil ditambahkan');
     }
@@ -96,14 +106,12 @@ class AdminHikingRouteController extends Controller
                     'location' => 'required|string|max:50',
                     'distance' => 'required',
                     'duration' => 'required',
-                    'elevation_gain' => 'required',
-                    'operating_hours' => 'nullable',
-                    'numb_of_posts' => 'nullable',
-                    'contact' => 'nullable',
+                    'operating_hours' => 'required',
+                    'numb_of_posts' => 'required',
+                    'contact' => 'required',
                     'fee' => 'nullable',
-                    'img' => 'nullable',
-                    'link' => 'nullable',
-                    'rules' => 'nullable',
+                    'lat' => 'required',
+                    'lng' => 'required',
                 ],
             );
 
@@ -112,48 +120,48 @@ class AdminHikingRouteController extends Controller
             }
 
             $hikingroute = HikingRoute::findOrFail($id);
+            $data = [
+                'mountain_id' => $request->mountain_id,
+                'name' => $request->name,
+                'status' => $request->status,
+                'difficulty' => $request->difficulty,
+                'location' => $request->location,
+                'distance' => $request->distance,
+                'duration' => $request->duration,
+                'operating_hours' => $request->operating_hours,
+                'numb_of_posts' => $request->numb_of_posts,
+                'contact' => $request->contact,
+                'fee' => $request->fee,
+                'rules' => $request->rules,
+                'lat' => $request->lat,
+                'lng' => $request->lng,
+            ];
+
             if ($request->hasFile('img')) {
+                // Menangani gambar jika ada
                 $img = $request->file('img');
                 $img->storeAs('public/hikingroutes', $img->hashName());
-
-                $path = str_replace('http://127.0.0.1:8000/storage/', '', $hikingroute->img);
-                Storage::delete('public/' . $path);
-
-                $hikingroute->update([
-                    'mountain_id' => $request->mountain_id,
-                    'name' => $request->name,
-                    'status' => $request->status,
-                    'difficulty' => $request->difficulty,
-                    'location' => $request->location,
-                    'distance' => $request->distance,
-                    'duration' => $request->duration,
-                    'elevation_gain' => $request->elevation_gain,
-                    'operating_hours' => $request->operating_hours,
-                    'numb_of_posts' => $request->numb_of_posts,
-                    'contact' => $request->contact,
-                    'fee' => $request->fee,
-                    'img' => $img->hashName(),
-                    'link' => $request->link,
-                    'rules' => $request->rules,
-                ]);
-            } else {
-                $hikingroute->update([
-                    'mountain_id' => $request->mountain_id,
-                    'name' => $request->name,
-                    'status' => $request->status,
-                    'difficulty' => $request->difficulty,
-                    'location' => $request->location,
-                    'distance' => $request->distance,
-                    'duration' => $request->duration,
-                    'elevation_gain' => $request->elevation_gain,
-                    'operating_hours' => $request->operating_hours,
-                    'numb_of_posts' => $request->numb_of_posts,
-                    'contact' => $request->contact,
-                    'fee' => $request->fee,
-                    'link' => $request->link,
-                    'rules' => $request->rules,
-                ]);
+                // Menghapus gambar lama jika ada
+                if ($hikingroute->img) {
+                    $path = str_replace('http://127.0.0.1:8000/storage/', '', $hikingroute->img);
+                    Storage::delete('public/' . $path);
+                }
+                $data['img'] = $img->hashName();
             }
+
+            if ($request->hasFile('file')) {
+                // Menangani file jika ada
+                $file = $request->file('file');
+                $file->storeAs('public/gpx', $file->hashName());
+                // Menghapus file lama jika ada
+                if ($hikingroute->file) {
+                    $pathFile = str_replace('http://127.0.0.1:8000/storage/', '', $hikingroute->file);
+                    Storage::delete('public/' . $pathFile);
+                }
+                $data['file'] = $file->hashName();
+            }
+
+            $hikingroute->update($data);
 
             return redirect()->route('admin.hikingroutes.index')->with('success', 'Data berhasil diupdate');
         } catch (\Exception $e) {
@@ -166,10 +174,19 @@ class AdminHikingRouteController extends Controller
         $hikingroute = HikingRoute::findOrFail($id);
 
         try {
-            if ($hikingroute->img) {
-                $path = str_replace('http://127.0.0.1:8000/storage/', '', $hikingroute->img);
-                Storage::delete('public/' . $path);
+            $filesToDelete = [
+                $hikingroute->img,
+                $hikingroute->file
+            ];
+
+            // Loop untuk menghapus file jika ada
+            foreach ($filesToDelete as $file) {
+                if ($file) {
+                    $path = str_replace('http://127.0.0.1:8000/storage/', '', $file);
+                    Storage::delete('public/' . $path);
+                }
             }
+
             $hikingroute->delete();
             return redirect()->route('admin.hikingroutes.index')->with('success', 'Data berhasil dihapus');
         } catch (\Exception $e) {
